Simplify/streamline.
