+++ src/main/java/edu/cnm/deepdive/salesforceoauthrelay/model/BaseCredential.java
@@ -58,4 +58,5 @@ public class BaseCredential {
   public void setPassword(String password) {
     this.password = password;
   }
+
 }
